MoteService updated..git status

+++ ms-noteservice/src/main/java/com/fundoonote/msnoteservice/service/NoteServiceImp.java
@@ -1,7 +1,11 @@
 package com.fundoonote.msnoteservice.service;
 
+import java.util.ArrayList;
 import java.util.Date;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -34,12 +38,13 @@ public class NoteServiceImp implements INoteService {
 
 	@Autowired
 	IJmsService jmsService;
-	
+
 	@Autowired
 	INotePrefDao notePrefDao;
 
-	/*@Autowired
-	IUserDao userDao;*/
+	/*
+	 * @Autowired IUserDao userDao;
+	 */
 
 	@Autowired
 	ICollaboratorDao collaboratorDao;
@@ -50,68 +55,131 @@ public class NoteServiceImp implements INoteService {
 	@Override
 	public void saveNote(NoteDto noteDto, String userId) throws NSException {
 
+		String imageUrl = null;
+		if (noteDto.getImage() != null) {
+			imageUrl = s3Service.saveImageToS3(noteDto.getNote().getNoteId(), noteDto.getImage());
+		}
 		Note note = noteDto.getNote();
 		note.setUserId(userId);
+		note.setCreatedDate(new Date());
+		note.setLastUpdated(new Date());
 		noteDao.save(note);
 		jmsService.addToQueue(note, OperationType.SAVE);
 		NotePreferences notePref = noteDto.getNotePreferences();
 		notePref.setNote(note);
-		Date date = new Date();
-		note.setCreatedDate(date);
-		note.setLastUpdated(date);
 		notePref.setUserId(userId);
 		notePrefDao.save(notePref);
 		jmsService.addToQueue(notePref, OperationType.SAVE);
 
+		if (noteDto.getNotePreferences().getLabels() != null) {
+			Set<Label> labels = noteDto.getNotePreferences().getLabels();
+			for (Label label : labels) {
+				label.setUserId(userId);
+				labelDao.save(label);
+			}
+			notePref.setLabels(labels);
+			notePref.setNote(note);
+			notePref.setUserId(userId);
+			notePrefDao.save(notePref);
+		}
+
+		if (noteDto.getCollaboratorId() != null) {
+			Set<String> collaborators = noteDto.getCollaboratorId();
+			for (String collaboratorid : collaborators) {
+				Collaboration collaboration = new Collaboration();
+				collaboration.setNote(note);
+				collaboration.setSharedById(userId);
+				collaboration.setSharedId(collaboratorid.toString());
+				collaboratorDao.save(collaboration);
+			}
+		}
 	}
 
 	@Override
 	public void updateNote(Note note, String userId) throws NSException {
 		Note oldNote = noteDao.getOne(note.getNoteId());
+		if (oldNote == null) {
+			throw new NSException(124, new Object[] { "note.getId()" });
+		}
+		if (oldNote.getUserId() == userId) {
+			throw new NSException(111, new Object[] { "update note " });
+		}
 		oldNote.setTitle(note.getTitle());
 		oldNote.setBody(note.getBody());
-		Date updatedDate = new Date();
-		oldNote.setLastUpdated(updatedDate);
+		oldNote.setLastUpdated(new Date());
 		noteDao.save(oldNote);
 		jmsService.addToQueue(oldNote, OperationType.UPDATE);
 	}
 
 	@Override
-	public void updatenotePref(NotePreferences notePref) {
+	public void updatenotePref(NotePreferences notePref, String loggedInUserId) throws NSException {
 
-		//noteDao.save(notePref);
+		Optional<NotePreferences> oldNotePreferences = notePrefDao.findById(notePref.getNotePreId());
+		if (!oldNotePreferences.isPresent()) {
+			throw new NSException(124, new Object[] { "notePref.getNotePreId()" });
+		}
+		if (!oldNotePreferences.get().getUserId().equals(loggedInUserId)) {
+			throw new NSException(111, new Object[] { "update note " });
+		}
+		notePrefDao.save(notePref);
+		jmsService.addToQueue(notePref, OperationType.UPDATE);
 	}
 
 	@Override
-	public void deleteNote(long noteId) throws NSException {
+	public void deleteNote(long noteId, String loggedInUserId) throws NSException {
 
+		Optional<Note> note = noteDao.findById(noteId);
+		if (!note.get().getUserId().equals(loggedInUserId)) {
+			throw new NSException(111, new Object[] { "delete note :-" });
+		}
 		noteDao.deleteById(noteId);
 		jmsService.addToQueue(noteId, OperationType.DELETE);
 	}
 
 	@Override
 	public List<NoteDto> getNotes(String loggedInUser) {
-
+        
+	    List<NoteDto> noteDTOs = new ArrayList<NoteDto>();
 		List<NotePreferences> notePreferences = notePrefDao.getAllNotePreferenceByUserId(loggedInUser);
 		List<NoteDto> result = notePreferences.stream().map(temp -> {
 			NoteDto noteDto = new NoteDto();
-			//noteDto.setNote(temp.getNote());
+			noteDto.setNote(temp.getNote());
+			noteDto.setCollaboratorId(getAllCollabUserByNote(temp.getNote().getNoteId()));
 			noteDto.setNotePreferences(temp);
 			return noteDto;
 		}).collect(Collectors.toList());
-
+		
 		return result;
 	}
 
+	private Set<String> getAllCollabUserByNote(long noteId) {
+		Note note = new Note();
+		note.setNoteId(noteId);
+		List<Collaboration> collaborators = collaboratorDao.getByNote(note);
+		Set<String> sharedUserIds = collaborators.stream().map(temp -> temp.getSharedId()).collect(Collectors.toSet());
+		return sharedUserIds;
+	}
+
 	@Override
-	public void saveLabel(Label label, String loggedInUserId) {
+	public void saveLabel(Label label, String loggedInUserId) throws NSException {
+		Label OldLabel = labelDao.findByNameAndUserId(label.getName(), loggedInUserId);
+		if (OldLabel != null) {
+			throw new NSException(115, new Object[] { label.getName() });
+		}
 		label.setUserId(loggedInUserId);
 		labelDao.save(label);
 	}
 
 	@Override
-	public void renameLabel(Label label, String loggedInUserId) {
+	public void renameLabel(Label label, String loggedInUserId) throws NSException {
 
+		Optional<Label> oldLabel = labelDao.findById(label.getLabelId());
+		if (!oldLabel.isPresent()) {
+			throw new NSException(114, new Object[] { "label.getLabelId()" });
+		}
+		if (!oldLabel.get().getUserId().equals(loggedInUserId)) {
+			throw new NSException(111, new Object[] { "Rename Label :-" });
+		}
 		Label labelFromDB = labelDao.getOne(label.getLabelId());
 		labelFromDB.setName(label.getName());
 		labelDao.save(labelFromDB);
@@ -124,16 +192,26 @@ public class NoteServiceImp implements INoteService {
 	}
 
 	@Override
-	public void deleteLabel(int labelId) {
+	public void deleteLabel(int labelId, String loggedInUserId) throws NSException {
+
+		Optional<Label> oldLabel = labelDao.findById(labelId);
+		if (!oldLabel.get().getUserId().equals(loggedInUserId)) {
+			throw new NSException(111, new Object[] { "Delete Label :-" });
+		}
 		labelDao.deleteById(labelId);
 	}
 
 	@Override
-	public void addLabel(long noteId, int labelId) {
+	public void addLabelToNote(long noteId, int labelId, String loggedInUserId) {
 
-		Note note = noteDao.getOne(noteId);
+		Note note = new Note();
+		note.setNoteId(noteId);
+		NotePreferences notePreferences = notePrefDao.getByNote(note);
 		Label label = labelDao.getOne(labelId);
-		//notePreference.
+		Set<Label> labels = new HashSet<Label>();
+		notePreferences.setLabels(labels);
+		notePrefDao.save(notePreferences);
+
 	}
 
 	@Override
@@ -164,7 +242,7 @@ public class NoteServiceImp implements INoteService {
 
 		Collaboration collaboration = new Collaboration();
 		Note note = noteDao.getOne(noteId);
-		//User sharedUser = userDao.findByEmail(sharingUserEmail);
+		// User sharedUser = userDao.findByEmail(sharingUserEmail);
 
 		collaboration.setNote(note);
 		collaboration.setSharedById(loggedInUserEmail);
@@ -173,42 +251,61 @@ public class NoteServiceImp implements INoteService {
 	}
 
 	@Override
-	public void removeCollaborator(String sharedUserId, long noteId) {
-		Note note = new Note();
-		note.setNoteId(noteId);
+	public void removeCollaborator(String sharedUserId, long noteId, String loggedInUserId) throws NSException {
+		Note note = noteDao.getOne(noteId);
+		if (note.getUserId() == loggedInUserId) {
+			throw new NSException(101, new Object[] { "" });
+		}
 		collaboratorDao.deleteByNoteAndSharedId(note, sharedUserId);
-
-	}
-	
-	@Override
-	public void trashOrRestore(long notePrefId, Status status, String loggedInUserId) {
-
-		NotePreferences notePref = notePrefDao.getOne(notePrefId);
-		notePref.setStatus(status);
-		notePref.setUserId(loggedInUserId);
-		notePrefDao.save(notePref);
-
+		note.setLastUpdated(new Date());
+		noteDao.save(note);
 	}
 
 	@Override
-	public void pinOrUnpin(long notePrefId, boolean isPinned, String loggedInUserId) {
+	public void trashOrRestore(long notePrefId, Status status, String loggedInUserId) throws NSException {
 
-		NotePreferences notePref = notePrefDao.getOne(notePrefId);
-		notePref.setPin(isPinned);
-		notePref.setUserId(loggedInUserId);
-		notePrefDao.save(notePref);
+		Optional<NotePreferences> notePref = notePrefDao.findById(notePrefId);
+		if (!notePref.get().getUserId().equals(loggedInUserId)) {
+			throw new NSException(111, new Object[] { "perform trash or Restore" });
+		}
+		if (status == null) {
+			throw new NSException(123, new Object[] { "perform trash or Restore" });
+		}
+		NotePreferences notePreferences = notePref.get();
+		notePreferences.setStatus(status);
+		notePreferences.setUserId(loggedInUserId);
+		notePrefDao.save(notePreferences);
+
 	}
 
 	@Override
-	public void archiveOrUnarchive(long notePrefId, Status status, String loggedInUserId) {
+	public void pinOrUnpin(long notePrefId, boolean isPinned, String loggedInUserId) throws NSException {
 
-		NotePreferences notePref = notePrefDao.getOne(notePrefId);
-		notePref.setStatus(status);
-		notePref.setUserId(loggedInUserId);
-		notePrefDao.save(notePref);
-
+		Optional<NotePreferences> notePref = notePrefDao.findById(notePrefId);
+		if (!notePref.get().getUserId().equals(loggedInUserId)) {
+			throw new NSException(111, new Object[] { "perform pin or Unpin" });
+		}
+		NotePreferences notePreferences = notePref.get();
+		notePreferences.setPin(isPinned);
+		notePreferences.setUserId(loggedInUserId);
+		notePrefDao.save(notePreferences);
 	}
 
+	@Override
+	public void archiveOrUnarchive(long notePrefId, Status status, String loggedInUserId) throws NSException {
 
+		Optional<NotePreferences> notePref = notePrefDao.findById(notePrefId);
+		if (!notePref.get().getUserId().equals(loggedInUserId)) {
+			throw new NSException(111, new Object[] { "perform archive or Unarchive" });
+		}
+		if (status == null) {
+			throw new NSException(123, new Object[] { "perform archive or Unarchive" });
+		}
+		NotePreferences notePreferences = notePref.get();
+		notePreferences.setStatus(status);
+		notePreferences.setUserId(loggedInUserId);
+		notePrefDao.save(notePreferences);
+
+	}
 
 }
